+++ test/utils_test.py
@@ -1,4 +1,7 @@
-""" unit test """
+"""
+    Apache License 2.0 Copyright (c) 2022 Pavel Bystrov
+    test user
+"""
 import unittest
 from webapp import create_app
 from webapp.business_logic import get_user_by_name
@@ -6,7 +9,6 @@ from webapp.business_logic import get_user_by_name
 
 class TestUser(unittest.TestCase):
     """ test user """
-
     def setUp(self):
         """ Инит TestUser """
         self.max_len = 500
