+++ webapp/utils/enrich_sunspots.py
@@ -1,9 +1,8 @@
-""" enrich sunspots utility """
-import pandas as pd
-import logging
+"""
+    Apache License 2.0 Copyright (c) 2022 Pavel Bystrov
+    enrich sunspots utility
+"""
 import numpy as np
-from numpy import hstack
-from optional import Optional
 from sklearn.ensemble import GradientBoostingClassifier, AdaBoostClassifier, \
     ExtraTreesClassifier, BaggingClassifier, RandomForestClassifier
 from sklearn.linear_model import LogisticRegression, RidgeClassifier
@@ -13,87 +12,11 @@ from sklearn.model_selection import GridSearchCV, StratifiedKFold, \
 from sklearn.neighbors import KNeighborsClassifier
 from sklearn.preprocessing import StandardScaler
 from sklearn.tree import DecisionTreeClassifier
-from webapp.utils.trends_util import rolling_mean, find_minimums
-
-
-def fill_values(data, ndx, func):
-    """ fill values from array :attr:`data` using function :attr:`func` """
-    if len(data) < 2:
-        return data
-    for index_ in ndx:
-        if index_ > len(data):
-            return np.array([])
-    vals = [np.ones((ndx[j + 1] - ndx[j])) * func(data[ndx[j]:ndx[j + 1]])
-            for j in range(len(ndx) - 1)]
-    result = hstack(vals)
-    return result
-
-
-def get_users_timeseries(csv_file: str = "data/hour_online.csv") -> Optional:
-    """
-       enrich users timeseries with 12 points moving average,
-       36 points moving averages and 128 points moving average
-    """
-    try:
-        times = pd.read_csv(csv_file)
-        # fill the first value of 34002 instead of NA
-        times["mean_12p"] = rolling_mean(times['Users'], 12).fillna(34002)
-        times["mean_36p"] = rolling_mean(times['Users'], 36).fillna(34002)
-        times["mean_128p"] = rolling_mean(times['Users'], 128).fillna(34002)
-        return Optional.of(times)
-    except FileNotFoundError:
-        message = f"File not found {csv_file}"
-        logging.error(message)
-        return Optional.empty()
-
-
-def get_enriched_dataframe(csv_file="data/sunspot_numbers.csv"):
-    """
-       enrich dataframe with 1y, 3y and 128 months moving averages and
-       with min, max and average number of sunspots
-    """
-    data = pd.read_csv(csv_file, delimiter=";")
-    trend = data['sunspots'].values
-    # calculate moving average
-    data["mean_1y"] = rolling_mean(data['sunspots'], 12)
-    data["mean_3y"] = rolling_mean(data['sunspots'], 36)
-    data["mean_12y"] = rolling_mean(data['sunspots'], 128)
-    # fill the first value of 96.7 instead of NA
-    data["mean_1y"] = data["mean_1y"].fillna(96.7)
-    data["mean_3y"] = data["mean_3y"].fillna(96.7)
-    data["mean_12y"] = data["mean_12y"].fillna(96.7)
-    # find minimums in trend using period = 128 months
-    mins = find_minimums(trend, 128)
-    # correction for short cycle after minimum #7 using period = 119 months
-    correction = find_minimums(trend[mins[7]:(mins[7] + 120)], 119)
-    # next cycle after minimum #7
-    cy8 = mins[7] + correction[1]
-    # correction for many zeroes at the end of minimum #5
-    cy6 = (mins[5] + mins[6]) // 2
-    # drop invalid minimums 6 and 9
-    indices = [0] + mins[:5] + [cy6, mins[7], cy8, mins[8]] + mins[10:] + \
-              [len(trend)]
-    # calculate min, max and average number of sunspots for solar cycles
-    min_ = fill_values(trend, indices, np.min)
-    max_ = fill_values(trend, indices, np.max)
-    avg = fill_values(trend, indices, np.mean)
-    data["sn_mean"] = pd.Series(avg.tolist())
-    data["sn_max"] = pd.Series(max_.tolist())
-    data["sn_min"] = pd.Series(min_.tolist())
-
-    y_max = hstack([np.zeros([indices[17]]),
-                    np.ones((indices[20] - indices[17])),
-                    np.zeros([indices[-1] - indices[20]])])
-    y_min = hstack([np.zeros([indices[5]]),
-                    np.ones((indices[8] - indices[5])),
-                    np.zeros([indices[-1] - indices[8]])])
-    data["y_min"] = pd.Series(y_min.tolist())
-    data["y_max"] = pd.Series(y_max.tolist())
-    return data
+from webapp.utils import dataframe_util
 
 
 def predict_using_cross_validation(clf, params, data, dframe):
-    """ predict with classifier using cross-validation with n=3"""
+    """ predict with classifier using cross-validation with n=3 """
     if clf is None or data is None or dframe is None:
         raise ValueError("Empty parameters")
     y_max = dframe["y_max"].values
@@ -136,13 +59,8 @@ def evaluate_classifier(clf, params, data_scaled, dframe):
     return predict_max, predict_min, max_, sunspots
 
 
-def get_results_for_best_classifier():
-    """ get results for best classifier """
-    dframe = get_enriched_dataframe()
-    times = dframe["year_float"].values
-    cols = ["sunspots", "observations", "mean_1y", "mean_3y", "mean_12y",
-            "sn_mean", "sn_max", "sn_min"]
-    data_scaled = StandardScaler().fit_transform(dframe[cols].values)
+def get_classifiers_and_params() -> list:
+    """ get classifiers and params """
     params_lr = {'C': np.linspace(8, 200, 20) / 10,
                  'class_weight': ["balanced", None]}
     params_rid = {'alpha': np.linspace(8, 200, 20) / 10,
@@ -160,10 +78,20 @@ def get_results_for_best_classifier():
         (BaggingClassifier(DecisionTreeClassifier()), estimators),
         (RandomForestClassifier(), params),
         (KNeighborsClassifier(), params_knn),
-        (ExtraTreesClassifier(), params),
-    ]
+        (ExtraTreesClassifier(), params)]
+    return classifiers
+
+
+def get_results_for_best_classifier():
+    """ get results for best classifier """
+    dframe = dataframe_util.get_enriched_dataframe()
+    times = dframe["year_float"].values
+    cols = ["sunspots", "observations", "mean_1y", "mean_3y", "mean_12y",
+            "sn_mean", "sn_max", "sn_min"]
+    data_scaled = StandardScaler().fit_transform(dframe[cols].values)
     max_score = 0.
     results = (ExtraTreesClassifier(), {})
+    classifiers = get_classifiers_and_params()
     for clf, parameters in classifiers:
         try:
             score, best_params = predict_using_cross_validation(clf,
